pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent-cli"
version = "0.1.0"
description = "CLI tool to generate high-quality prompts for tests, docs, and design."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Yaroslav Tanko" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Environment :: Console",
    "Topic :: Software Development :: Code Generators"
]
dependencies = [
    "typer>=0.12,<1.0",
    "jinja2>=3.1,<4.0",
    "pathspec>=0.12,<1.0",
    "pyyaml>=6.0,<7.0",
    "rich>=13.0,<14.0",
    # LangChain + OpenAI-compatible client for LM Studio
    "langchain>=0.2.12,<0.3.0",
    "langchain-openai>=0.1.23,<0.2.0",
]

[project.urls]
Homepage = "https://example.com/ai-agent-cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["ai_agent*"]

[project.scripts]
ai-agent = "ai_agent.cli:app"

